[#47] Refactor plug-ins to support newest changes in emuLib

+++ plugins/cpu/8080-cpu/src/test/java/net/sf/emustudio/intel8080/impl/InstructionsTest.java
@@ -122,7 +122,7 @@ public abstract class InstructionsTest {
     protected void stepAndCheckPCandSPandMemory(int PC, int SP, int memValue) {
         stepAndCheckPC(PC);
         assertEquals(SP, cpu.getEngine().SP);
-        assertEquals(memValue, memoryStub.readWord(cpu.getEngine().SP));
+        assertEquals(memValue, memoryStub.readWord(cpu.getEngine().SP).intValue());
     }
 
     protected void stepAndCheckPCandSP(int PC, int SP) {
